requirements.txt
django>=4.2,<5.0
psycopg2
Pillow
django-filter
django-tables2
django-crispy-forms
crispy-bootstrap5
django-dbbackup 
celery
kombu
redis
